requirements.txt
Django==5.2.5
mysqlclient==2.2.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
celery==5.4.0
redis==5.0.7
websocket-client==1.8.0
python-dotenv==1.0.0
docker==7.1.0
beautifulsoup4==4.12.3
pydantic==2.10.3
langchain==0.3.0
langchain-openai==0.2.0
openai==1.55.3
yfinance==0.2.40
